New Partitioning and Column Usage Reports
3 new reports to help understand the oppotunities for partitioning in a
database
* Section 2b: Column Usage of 100 Largest Tables: For 100 largest tables
by number of blocks (as recorded on DBA_TABLES), show the column usage
where known, and either indicate that the table is not partitioned, or
show the partition/subpartition key columns.  The idea here is to look
for tables that are candidates for partitioning, possibly with a view to
compression/archiving.
* Section 3c: Table Partitioning
Report of one row per partitioned table across the database, showing
sizes, numbers of partition/subpartitions, how many of them are
compressed at what level, partitioning key columns
* Section 3c: Partitioning Keys and Column Usage Statistics
Report of partition tables, showing a row for each column in column
usage or that is a partition/subpartition key column.  So you can see
whether the existing partitioning is effective in that it matches the
way the applcation uses the table.
Not compleyely dissimilar to the new report in 2b, but that shows tables
whether partitioned or now.