Columns with Histograms in Extended Statistics
New report in section 3c.  Check for columns with histograms that are
either
i) part of an extended statistics (ie on a group of columns) where there
is no histogram (its is not an extended histogram)
ii) part of a composite key index where there is no extended statistics
In either case the presence of the histogram on the consituent column
prevents the optimiser using the number of distinct keys in the extended
statistics.  The workaround is to build an extended histogram on the
column group.