Corrections to average wait time for event histogram buckets
The duration of the histogram bucket in some summary reports (was being calculated as half the difference in the upper limit for the current and previous lower bucket.  Therefore, these reports are returning incorrect values.  The average should be calculated as the average of the two, so it should be half the sum.  Where there isn't a previous lower bucket it should be half the upper limit on the bucket.  The lag function returns null, so NVL(...,0) has been added.
WAIT_CLASS_ID has been removed from the partition by clauses in the analytic functions because they already contain EVENT_ID.