Change to use all event histogram buckets but cap the max value
Adjustment to wait time calculation.  It now uses all the buckets, but caps the bucket time for the highest bucket.  All the other buckets are twice the duration of the preceding bucket, the maximum bucket is 1.8E16ms, effectively infinte.  So this value is capped at 1.5x the bucket below so that it follows the progression (it works out at 6291456).
It probably won't make any difference to any of the reports because only Idle waits seem to record waits in this bucket.