7b remodel for fixes and performance
The 7b section was remodel to fix several issues:
- The ora-32034 when materializing the WITH clauses
- The syntactic errors depending on RDBMS versions and multi-tenant architecture
- The sqld360 generated files failed to be added to the edb360 main file
- Generating SQLs for duplicate SQLIDs, for PL/SQLs and other SQLs that the SQLD360 is not designed to collect info from
- SQL texts sometimes empty

The complexity of the driving query makes the maintenance and implementation of criteria more difficult
and an all-or-nothing result, meaning if one section failed for whatever reason no SQLD360 or planx is generated.

The driving way replaced by 3 materializations of the 3 criteria used to identify SQLs to generate the SQLD360 and Planx.
The materializations use PLAN_TABLE to hold the rows.
The PLAN_TABLE is not going to hold more rows than the sum of &&edb360_conf_top_sql + &&edb360_conf_top_cur + &&edb360_conf_top_sig .

The top_sql materialization looks for INSERT,UPDATES,DELETS,SELECT, CTAS AND CREATE MVS
that have at least 10 minutes of accumulated DB time.
excludes the SQLS coming from edb360,sqld360,MMON process and auto-stats job.

The top_not_shared materialization looks for SQLs that have a high version count larger than 100 in the library cache.
The dba_hist_sqlstat has a record of the SQLs having high version count but the information from v$sql_shared_cursors is
not persistent in the AWR as of 19c so that is why on this materialization looks for the SQLs in the library cache.
It is expensive but at this time there is no other way.
Excludes the SQLs from top_sql because the sqld360 is already captured for them.

The top_signature materialization looks for SQLs that have high DB time grouped by force_matching_signature.
No SQLs is excluded like in "top_sql" because if it is possible that any group of SQLs including our own to generate multiple literal SQLs
that cause a high impact we want to know.
Excludes the SQLs from top_sql and top_non_shared because the sqld360 is already captured for them.

The materialized rows then are matched to their respective pdb_name and username.
The rows are matched with their SQL text using multiple sources without considering the corresponding pdb because the SQLID is hash of the SQL Text so it is the same
regardless that the SQL is executed in multiple PDBs.