6f,6k,6l Improvements
6f Top Modules and Actions
Improved the Operations profile to show "Other Modules" and "No Modules" to have a better comparision with the Top Modules.

6k Top PHV
This section has the objective to identify SQLs (plural) that shared the same plan that collectively may be high consumers
but individualy mey not show in 6c Top SQLs

6l Top Signature
This section has the objective to identify SQLs (Plural) that shared the same force_matching_signature that collectively
may be high consumers but individually may not show in 6c Top SQLs.