4f Repurpose to AAS by Class and Top Events
4f section is divided in 2:
- Graphs to show the Average Active Sessions by instance for each class
- Graphs to show the Average Active Sessions by Top wait events in each class

Section 5b shows similar information that the first section but the difference is that they use a different source for the calculations.
5b uses active session history where each row represents 10 seconds of DB time so each event or CPU is assumed to have taken 10 seconds regardless if it took less.
4f uses wait event histograms which are much more granular in the exact DB time used by snap as each snap has several buckets that hold how many waits happened in each bucket
that starts from >1ms and up to seconds.
This means that 5b is granular across snap time (Horizontally) and 4f is granular across magnitude (Vertically) so they both will show similar-looking graphs but not identical and the true
representation of the workload is between the two.

The second set of graphs show the Top Wait events that contribute to the AAS in the same class so the shape of the graphs will look the same as the first set but
the contents will be subdivided by the contributors. This is the intermediate step to the 4g section that shows each of bucket's contribution to AAS for just one event at a time.
So sections 4f and 4g effectively show a Top-Down break down of the AAS at 3 different levels, By instance for each class, by events in each Class, By bucket in each event.
4f section can be further improved to break down by events in each Class in each instance.

This section is using an experimental method that places the summary of the data in the plan_table to make the queries faster but it might fail in a standby env.

There is no consideration yet to breakdown by PDB. this is a later project.